dsads

+++ run_tnsm_mif.py
@@ -202,7 +202,7 @@ def get_corrected_filename(dataset, mif, seed, threshold, window, full=True):
     else:
         filename = "{}_pif-{:0>2d}.corrected_threshold-{:0>2d}_window-{:0>2d}_epochs-{:0>4d}".format(
             get_original_unzip_filename(dataset, False),
-            mif*100, threshold, window, NUM_EPOCHS)
+            int(mif*100), threshold, window, NUM_EPOCHS)
 
     if full:
         filename = "{}/{}".format(PATH_CORRECTED, filename)
